requirements.txt
openai
tweepy
requests
gTTS
apscheduler
Pillow
